Successfully converting to RREF.

+++ src/matrix_elements/row.c
@@ -1,1 +1,1 @@
-typedef int Row[COLUMN_COUNT];
+typedef float Row[COLUMN_COUNT];
